+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - Fix broken functionality for shared items in OneDrive/Sharepoint. In particular, this should allow using the MS365 backend with the pins package (#149/#129).
 - The `list_shared_items`/`list_shared_files` method for drives now always returns a list of drive item objects, rather than a data frame. If the `info` argument is supplied with a value other than "items", a warning is issued.
+- Add folder download functionality for `ms_drive_item$download()` to download all the files in a folder, possibly recursively and in parallel. There is also a corresponding `ms_drive$download_folder()` method.
 
 # Microsoft365R 2.3.4